Misc. cleanup for DECTRIS support and new APIs
* Implement `LiveContext.connect`, for now only supporting DECTRIS
  detectors
* `LiveContext.prepare_acquisition`: if no trigger function is given,
  construct a no-op default.
* Fix return type of `DetectorConnection.wait_for_acquisition`
* Split up DECTRIS support into multiple files
* Remove some unused code
* Add test cases for passive mode
