+++ src/libertem_live/detectors/base/__init__.py
@@ -1,0 +1,4 @@
+from .connection import DetectorConnection
+from .acquisition import AcquisitionMixin
+
+__all__ = ["DetectorConnection", "AcquisitionMixin"]
