+++ src/libertem_live/detectors/dectris/common.py
@@ -1,0 +1,23 @@
+from typing import NamedTuple, Union
+from typing_extensions import Literal
+
+
+TriggerMode = Union[
+    Literal['exte'],  # acquire one image for each trigger, `ntrigger` times
+    Literal['exts'],  # acquire series of `nimages` with a single trigger
+    Literal['ints'],  # internal software triggering
+    Literal['inte'],  # internal software enable -> one image for each soft-trigger
+]
+
+
+class AcquisitionParams(NamedTuple):
+    sequence_id: int
+    nimages: int
+
+
+class DetectorConfig(NamedTuple):
+    x_pixels_in_detector: int
+    y_pixels_in_detector: int
+    bit_depth: int
+
+
